RetinaKAN/requirements.txt
# RetinaKAN – core dependencies (Python 3.8+)
# Install PyTorch first for your CUDA version, e.g.:
#   pip install torch torchvision torchaudio

# Training & evaluation
torch
torchvision
numpy>=1.20,<2
albumentations>=1.0
opencv-python
tqdm
PyYAML

# Model (archs.py)
timm>=0.6

# Metrics & plots (train.py, val.py, test_eval.py, metrics.py)
scikit-learn>=1.0
matplotlib>=3.3
medpy>=0.4

# Optional: used by config.py if you use it
yacs>=0.1.8

# Optional: image I/O (albumentations/cv2 use internally)
pillow

# --- Unused by main scripts (safe to remove if you only run train/val/test_eval) ---
# addict
# tensorboardX
# perceptual
# scikit-image
# tifffile
# typing-extensions  # often pulled in by timm
